+++ source/release-notes.md
@@ -941,3 +941,4 @@
 
 13.10.2019
 * monitoring: improved performance for transaction status; include the actual SQL code in the collected data / alerts sent
+* option to add/remove databases from a project using name filters